requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.1
apispec==5.2.2
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
avro==1.10.2
avro-gen3==0.7.10
Babel==2.12.1
blinker==1.6.2
cached-property==1.5.2
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-default-group==1.2.2
click-spinner==0.1.10
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.2
Deprecated==1.2.14
dill==0.3.7
dnspython==2.4.0
docker==6.1.3
docutils==0.20.1
email-validator==1.3.1
entrypoints==0.4
exceptiongroup==1.1.2
expandvars==0.9.0
frozenlist==1.4.0
google-re2==1.0
google-cloud-storage==2.11.0
graphviz==0.20.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
humanfriendly==10.0
idna==3.4
ijson==3.2.3
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonref==1.1.0
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
mixpanel==4.10.0
multidict==6.0.4
mypy-extensions==1.0.0
ordered-set==4.1.0
packaging==23.1
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.2.0
prison==0.2.1
progressbar2==4.2.0
psutil==5.9.5
pycparser==2.21
pydantic==1.10.12
Pygments==2.15.1
PyJWT==2.8.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-slugify==8.0.1
python-utils==3.7.0
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
ratelimiter==1.2.0.post0
referencing==0.30.0
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.4.2
rich-argparse==1.2.0
rpds-py==0.9.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
toml==0.10.2
typing-inspect==0.9.0
typing_extensions==4.5.0
tzlocal==5.0.1
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.16
websocket-client==1.6.1
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.16.2
spotipy==2.23.0
google~=3.0.0
pandas~=2.1.1